% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.scoringutils}
\alias{.scoringutils}
\title{Format `ggoutbreak` data for `scoringutils`}
\usage{
.scoringutils(prediction, raw, changes, ip)
}
\arguments{
\item{prediction}{the output from ggoutbreak `incidence` and `rt` predictions + maybe grouping}

\item{raw}{the raw incidence dataframe (`time` and `count`) + same grouping as prediction}

\item{changes}{a change point or time series for $R_t$ with columns `t` and `R_t`}

\item{ip}{the infectivity profile(s) with columns `boot`,`time`, and `probability`:}
}
\value{
a long format suitable for `scoringutils::score`
}
\description{
Format `ggoutbreak` data for `scoringutils`
}
