% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-test-utils.R
\name{.test_branching_process}
\alias{.test_branching_process}
\title{Generate a line list from a branching process model parameterised by
   reproduction number}
\usage{
.test_branching_process(
  changes = tibble::tibble(t = c(0, 40), R_t = c(2.5, 0.8)),
  kappa = 1,
  ip = ggoutbreak::covid_infectivity_profile,
  max_time = 80,
  initial = 30,
  seed = Sys.time(),
  summarise = TRUE,
  ...
)
}
\arguments{
\item{changes}{a dataframe containing a `t` time column and `R_t` reproduction
number parameter.}

\item{kappa}{a dispersion parameter controlling the likelihood of individual
super-spreading. This must be between 1 and Inf with 1 being standard poisson
dispersion and large values representing over dispersion.}

\item{ip}{a data frame with `time`, and `probability` columns}

\item{max_time}{maximum duration of simulation}

\item{initial}{initial outbreak size}

\item{seed}{random seen}

\item{summarise}{if `FALSE` return the line list of one entry per person}

\item{...}{
  Arguments passed on to \code{\link[=.summarise_with_ascertainment]{.summarise_with_ascertainment}}
  \describe{
    \item{\code{p.asc}}{the background probability of a case being detected}
    \item{\code{kappa.asc}}{a dispersion parameter controlling the day to day
variability of ascertainment between 0 (no dispersion) and Inf (maximum)}
  }}
}
\value{
either a line list of cases, with individual ids, infection times,
   and infector ids or a summary case count time series,
}
\description{
Generate a line list from a branching process model parameterised by
   reproduction number
}
\examples{
if (interactive()) {

tmp = .test_branching_process(
  changes = tibble::tibble(t = c(0,20,40,60,80,110), R_t = c(1.8,1.5,0.9,1.5,0.8,1.2)),
  kappa = 2,
  max_time = 120,
  p.asc= 0.8,
  kappa.asc = 1.01,
  seed = 100
)

mean(tmp$dispersion)
ggplot2::ggplot(tmp, ggplot2::aes(x=time,y=count))+ggplot2::geom_point()

}

}
\keyword{internal}
